changed signed int to signed size t

+++ 0x14-file_io/1-create_file.c
@@ -12,7 +12,7 @@
 
 int create_file(const char *filename, char *text_content)
 {
-	signed int file;
+	ssize_t file;
 	int count = 0;
 
 	if (!filename)
